chore: update project files and fix various code issues
- Update Visual Studio platform toolset from v142 to v143 across multiple projects
- Fix signed/unsigned comparison warnings by using correct loop variable types
- Remove unused exception variable name in catch block
- Fix potential out-of-bounds access in UI mix expansion
- Correct hex color value in event time display
- Convert file encoding from UTF-8 BOM to UTF-8
- Translate Korean comments to Korean text (maintaining original language)
- Ensure proper file endings and formatting

+++ Main5.2/source/CBInterface.cpp
@@ -597,7 +597,7 @@ bool CB_Interface::DrawSelectBar(int X, int Y, int W, int H, int* SelectNum, std
 	{
 		int CountSelectBar = 1;
 		gInterface->DrawBarForm(X + 15, Y + 41 + (15), W - 25, (H - 5) * (Data.size()), 0.0, 0.0, 0.0, 0.8);
-		for (int i = 0; i < Data.size(); i++)
+		for (size_t i = 0; i < Data.size(); i++)
 		{
 			DWORD SelectHover = 0x0;
 			if (SEASON3B::CheckMouseIn(X + 15, Y + 41 + (15 * CountSelectBar), W + 20, 15))
